Remove need for bump seeds from mint wrapper
