+++ programs/quarry-mint-wrapper/src/instructions/mod.rs
@@ -1,0 +1,4 @@
+pub mod new_minter;
+pub mod new_wrapper;
+pub use new_minter::*;
+pub use new_wrapper::*;
