+++ programs/quarry-mint-wrapper/src/lib.rs
@@ -1,6 +1,7 @@
 //! Proxy program for interacting with the token mint.
 #![deny(rustdoc::all)]
 #![allow(rustdoc::missing_doc_code_examples)]
+#![allow(deprecated)]
 
 #[macro_use]
 mod macros;
@@ -11,6 +12,9 @@ use anchor_spl::token::{self, Mint, TokenAccount};
 use vipers::prelude::*;
 
 mod account_validators;
+mod instructions;
+
+use instructions::*;
 
 declare_id!("QMWoBmAyJLAsA1Lh9ugMTw2gciTihncciphzdNzdZYV");
 
@@ -29,33 +33,23 @@ solana_security_txt::security_txt! {
 pub mod quarry_mint_wrapper {
     use super::*;
 
-    /// --------------------------------
-    /// [MintWrapper] instructions
-    /// --------------------------------
+    // --------------------------------
+    // [MintWrapper] instructions
+    // --------------------------------
 
     /// Creates a new [MintWrapper].
+    #[deprecated(since = "5.0.0", note = "Use `new_wrapper_v2` instead.")]
     #[access_control(ctx.accounts.validate())]
     pub fn new_wrapper(ctx: Context<NewWrapper>, _bump: u8, hard_cap: u64) -> Result<()> {
-        let mint_wrapper = &mut ctx.accounts.mint_wrapper;
-        mint_wrapper.base = ctx.accounts.base.key();
-        mint_wrapper.bump = unwrap_bump!(ctx, "mint_wrapper");
-        mint_wrapper.hard_cap = hard_cap;
-        mint_wrapper.admin = ctx.accounts.admin.key();
-        mint_wrapper.pending_admin = Pubkey::default();
-        mint_wrapper.token_mint = ctx.accounts.token_mint.key();
-        mint_wrapper.num_minters = 0;
+        instructions::new_wrapper::handler(ctx, hard_cap)
+    }
 
-        mint_wrapper.total_allowance = 0;
-        mint_wrapper.total_minted = 0;
-
-        emit!(NewMintWrapperEvent {
-            mint_wrapper: mint_wrapper.key(),
-            hard_cap,
-            admin: ctx.accounts.admin.key(),
-            token_mint: ctx.accounts.token_mint.key()
-        });
-
-        Ok(())
+    /// Creates a new [MintWrapper].
+    ///
+    /// The V2 variant removes the need for supplying the bump.
+    #[access_control(ctx.accounts.validate())]
+    pub fn new_wrapper_v2(ctx: Context<NewWrapper>, hard_cap: u64) -> Result<()> {
+        instructions::new_wrapper::handler(ctx, hard_cap)
     }
 
     /// Transfers admin to another account.
@@ -93,31 +87,18 @@ pub mod quarry_mint_wrapper {
     /// --------------------------------
 
     /// Creates a new [Minter].
+    #[deprecated(since = "5.0.0", note = "Use `new_minter_v2` instead.")]
     #[access_control(ctx.accounts.validate())]
     pub fn new_minter(ctx: Context<NewMinter>, _bump: u8) -> Result<()> {
-        let minter = &mut ctx.accounts.minter;
+        instructions::new_minter::handler(ctx)
+    }
 
-        minter.mint_wrapper = ctx.accounts.auth.mint_wrapper.key();
-        minter.minter_authority = ctx.accounts.new_minter_authority.key();
-        minter.bump = unwrap_bump!(ctx, "minter");
-
-        let index = ctx.accounts.auth.mint_wrapper.num_minters;
-        minter.index = index;
-
-        // update num minters
-        let mint_wrapper = &mut ctx.accounts.auth.mint_wrapper;
-        mint_wrapper.num_minters = unwrap_int!(index.checked_add(1));
-
-        minter.allowance = 0;
-        minter.total_minted = 0;
-
-        emit!(NewMinterEvent {
-            mint_wrapper: minter.mint_wrapper,
-            minter: minter.key(),
-            index: minter.index,
-            minter_authority: minter.minter_authority,
-        });
-        Ok(())
+    /// Creates a new [Minter].
+    ///
+    /// The V2 variant removes the need for supplying the bump.
+    #[access_control(ctx.accounts.validate())]
+    pub fn new_minter_v2(ctx: Context<NewMinter>) -> Result<()> {
+        instructions::new_minter::handler(ctx)
     }
 
     /// Updates a [Minter]'s allowance.
@@ -187,75 +168,9 @@ pub mod quarry_mint_wrapper {
     }
 }
 
-/// --------------------------------
-/// Instructions
-/// --------------------------------
-
-#[derive(Accounts)]
-pub struct NewWrapper<'info> {
-    /// Base account.
-    pub base: Signer<'info>,
-
-    #[account(
-        init,
-        seeds = [
-            b"MintWrapper".as_ref(),
-            base.key().to_bytes().as_ref()
-        ],
-        bump,
-        payer = payer,
-        space = 8 + MintWrapper::LEN
-    )]
-    pub mint_wrapper: Account<'info, MintWrapper>,
-
-    /// CHECK: Admin-to-be of the [MintWrapper].
-    pub admin: UncheckedAccount<'info>,
-
-    /// Token mint to mint.
-    pub token_mint: Account<'info, Mint>,
-
-    /// Token program.
-    pub token_program: Program<'info, Token>,
-
-    /// Payer.
-    #[account(mut)]
-    pub payer: Signer<'info>,
-
-    /// System program.
-    pub system_program: Program<'info, System>,
-}
-
-/// Adds a minter.
-#[derive(Accounts)]
-pub struct NewMinter<'info> {
-    /// Owner of the [MintWrapper].
-    pub auth: OnlyAdmin<'info>,
-
-    /// Account to authorize as a minter.
-    /// CHECK: Can be any Solana account.
-    pub new_minter_authority: UncheckedAccount<'info>,
-
-    /// Information about the minter.
-    #[account(
-        init,
-        seeds = [
-            b"MintWrapperMinter".as_ref(),
-            auth.mint_wrapper.key().to_bytes().as_ref(),
-            new_minter_authority.key().to_bytes().as_ref()
-        ],
-        bump,
-        payer = payer,
-        space = 8 + Minter::LEN
-    )]
-    pub minter: Account<'info, Minter>,
-
-    /// Payer for creating the minter.
-    #[account(mut)]
-    pub payer: Signer<'info>,
-
-    /// System program.
-    pub system_program: Program<'info, System>,
-}
+// --------------------------------
+// Instructions
+// --------------------------------
 
 /// Updates a minter.
 #[derive(Accounts)]
